SR - Add application launchers and minor adjustments to the emacs grammar
Application launchers were added to the desktop environment grammar. Additionally, a note was added
to the README file of the caster/kaldi toolchain that explains why the StartApp Dragonfly action
cannot be used when running Dragonfly from inside a container.

Finally, a new `fill paragraph` command was added to the emacs.